--- src/test/java/Homework23.java
@@ -1,22 +1,0 @@
-import org.testng.Assert;
-import org.testng.annotations.Test;
-import pagefactory.RegistrationPage;
-
-public class Homework23 extends BaseTest{
-    @Test
-    public void registrationNavigation() {
-
-        //RegistrationPage registrationPage = new RegistrationPage(driver);
-        RegistrationPage registrationPage = new RegistrationPage(getDriver());
-
-        //Navigate to the page
-        registrationPage.navigateToMainPage();
-
-        //Click registration link
-        registrationPage.clickRegistrationLink();
-
-        //Expected Results
-        String registrationUrl = "https://qa.koel.app/registration";
-        Assert.assertEquals(registrationPage.getCurrentUrl(), registrationUrl);
-    }
-}
