+++ src/test/java/RegistrationTest.java
@@ -1,12 +1,8 @@
 import org.testng.Assert;
 import org.testng.annotations.Test;
 import pagefactory.RegistrationPage;
-import io.github.bonigarcia.wdm.config.WebDriverManagerException;
-import  io.github.bonigarcia.wdm.config.WebDriverManagerException;
 
-
-
-public class Homework16 extends BaseTest {
+public class RegistrationTest extends BaseTest {
     @Test
     public void registrationNavigation() {
