+++ src/test/java/SongTest.java
@@ -1,29 +1,33 @@
 import org.testng.Assert;
 import org.testng.annotations.Test;
+import pagefactory.AllSongsPage;
 import pagefactory.HomePage;
 import pagefactory.LoginPage;
 
-public class Homework22 extends BaseTest{
+public class SongTest extends BaseTest {
 
     @Test
-    public void renamePlaylist(){
+    public void playSong() {
 
         //LoginPage loginPage = new LoginPage(driver);
         //HomePage homePage = new HomePage(driver);
+        //AllSongsPage allSongs = new AllSongsPage(driver);
         LoginPage loginPage = new LoginPage(getDriver());
         HomePage homePage = new HomePage(getDriver());
+        AllSongsPage allSongs = new AllSongsPage(getDriver());
 
         loginPage.provideEmail("jennifer.de.jesus@testpro.io").providePassword("FCVlLOni").clickSubmit();
+
         Assert.assertTrue(homePage.getUserAvatar().isDisplayed());
 
-        String playlistId = "Jennys fourth Playlist";
-        String newPlaylistName = "Edited Jennys fourth Playlist";
+        homePage.chooseAllSongsList();
 
-        homePage.openPlaylist2(playlistId);
-        homePage.renamePlaylist(newPlaylistName);
+        allSongs.contextClickFirstSong();
 
-        String updatedPlaylistMsg = "Updated playlist \"Edited Jennys fourth Playlist.\"";
-        Assert.assertEquals(homePage.getRenameMessage(), updatedPlaylistMsg);
+        allSongs.hoverMediaPlayer();
+
+        allSongs.clickPlay();
+
+        Assert.assertTrue(allSongs.isSongPlaying());
     }
 }
-
