+++ src/test/resources/SongTest.xml
@@ -1,12 +1,12 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
 
-<suite name="Homework18">
+<suite name="SongTest">
 
     <parameter name="BaseURL" value="https://qa.koel.app/"/>
     <test name="Example Test" preserve-order="false">
         <classes>
-            <class name="Homework18">
+            <class name="SongTest">
                 <methods>
                     <include name="playSong"/>
                 </methods>
