+++ src/test/resources/TestNG.xml
@@ -5,15 +5,10 @@
     <parameter name="BaseURL" value="https://qa.koel.app/"/>
     <test name="Example Test" preserve-order="false">
         <classes>
-            <class name="LoginTests"/>
             <class name="BaseTest"/>
-            <class name="Homework16"/>
-            <class name="Homework17"/>
-            <class name="Homework19"/>
-            <class name="Homework20"/>
-            <class name="Homework21"/>
-            <class name="Homework22"/>
-            <class name="Homework23"/>
+            <class name="RegistrationTest"/>
+            <class name="LoginTests"/>
+            <class name="PlaylistTests"/>
         </classes>
     </test>
 </suite>
